trying some test things
A bunch of this stuff is just rubbish. I'm playing around with things
trying to get a feel for how this should work. I'm hoping to read a
config file from an envvar, or just fallback to one that already exists.
And if that doesn't work then we should skip the tests, probably?